+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Improved column width calculation
+
 ## [0.4.10] - 2026-04-17
 
 ### Added